cli/pyproject.toml
[tool.poetry]
name = "starfish-cli"
version = "0.1.0"
description = "CLI for Starfish-FL — used by humans and AI agents"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.10"
typer = {version = "0.9.0", extras = ["all"]}
requests = "^2.31"
python-dotenv = "^1.0"
rich = "^13.0"
click = "8.1.7"

[tool.poetry.scripts]
starfish = "starfish_cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"